+++ src/main/java/com/BoomECI/persistencia/ConsejeroDAO.java
@@ -14,8 +14,8 @@ import java.util.List;
  */
 public interface ConsejeroDAO {    
     
-    public List<SolicitudCancelacion> loadSolicitudesTramitadas(int id);
-    public List<SolicitudCancelacion> loadSolicitudesNoTramitadas(int id);    
+    public List<SolicitudCancelacion> loadSolicitudesTramitadas(long id);
+    public List<SolicitudCancelacion> loadSolicitudesNoTramitadas(long id);    
     public void cambiarEstadoSolicitud(int id_soli, String estado);
     public void agregarComentarioConsejero(int id_soli, String comentario);
     public void cambiarElAvalDeConsejero(int id_soli, boolean aval);
